♻️ 月、週、日の集計処理を共通ユーリティ関数としてまとめた
