+++ packages/fukui-terminal/src/App.tsx
@@ -1,12 +1,12 @@
 import { useEffect, useState } from "react";
-import { AggregatedData, formatDate, getRawData, TOTAL_COUNT_KEY } from "@fukui-kanko/shared";
+import { AggregatedData, getRawData } from "@fukui-kanko/shared";
 import {
   Graph,
   MonthRangePicker,
   RangeSelector,
   TypeSelect,
 } from "@fukui-kanko/shared/components/parts";
-import * as holidayJP from "@holiday-jp/holiday_jp";
+import { aggregateDaily, aggregateMonthly, aggregateWeekly } from "@fukui-kanko/shared/utils";
 
 function App() {
   // 開発環境かどうかを判定
@@ -50,107 +50,14 @@ function App() {
     if (type === "month" && startMonth && endMonth) {
       // 月末を取得
       const end = new Date(endMonth.getFullYear(), endMonth.getMonth() + 1, 0);
-      // 範囲でフィルタ
-      filtered = filtered.filter((row) => {
-        const date = new Date(row["aggregate from"]);
-        return date >= startMonth && date <= end;
-      });
-
-      // 月ごとに集計
-      const monthlyMap = new Map<string, AggregatedData>();
-      filtered.forEach((row) => {
-        const date = new Date(row["aggregate from"]);
-        const monthKey = `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, "0")}`;
-        if (!monthlyMap.has(monthKey)) {
-          monthlyMap.set(monthKey, {
-            ...row,
-            aggregateFrom: monthKey,
-            aggregateTo: monthKey,
-            totalCount: Number(row[TOTAL_COUNT_KEY]),
-          });
-        } else {
-          const prev = monthlyMap.get(monthKey)!;
-          monthlyMap.set(monthKey, {
-            ...prev,
-            totalCount: Number(prev.totalCount) + Number(row[TOTAL_COUNT_KEY]),
-          });
-        }
-      });
-      setFilteredData(Array.from(monthlyMap.values()));
-      return;
+      filtered = aggregateMonthly(filtered, startMonth, end);
     }
 
     if (type === "week" && startWeekRange && endWeekRange) {
-      // 週の範囲でフィルタ
-      filtered = filtered.filter((row) => {
-        const date = new Date(row["aggregate from"]);
-        return date >= startWeekRange.from && date <= endWeekRange.to;
-      });
-
-      const weeklyAggregated: AggregatedData[] = [];
-      let i = 0;
-      let isFirstWeek = true;
-      while (i < filtered.length) {
-        let weekRows: AggregatedData[] = [];
-        if (isFirstWeek) {
-          // 最初の週だけ他の週と範囲が異なる場合があるためfilterで抽出(例：2024/10/17からの週)
-          weekRows = filtered.filter((row) => {
-            const d = new Date(row["aggregate from"]);
-            return d >= startWeekRange.from && d <= startWeekRange.to;
-          });
-          // 件数分インデックスを進める
-          i += weekRows.length;
-          isFirstWeek = false;
-        } else {
-          // 以降は7日ごと
-          weekRows = filtered.slice(i, i + 7);
-          i += 7;
-        }
-        const total = weekRows.reduce((sum, row) => sum + Number(row[TOTAL_COUNT_KEY]), 0);
-        weeklyAggregated.push({
-          ...weekRows[0],
-          aggregateFrom: `${formatDate(new Date(weekRows[0]["aggregate from"]), "-")}〜`,
-          aggregateTo: formatDate(new Date(weekRows[weekRows.length - 1]["aggregate from"]), "-"),
-          totalCount: total,
-        });
-      }
-      setFilteredData(weeklyAggregated);
-      return;
+      filtered = aggregateWeekly(filtered, startWeekRange, endWeekRange);
     }
     if (type === "day" && startDate && endDate) {
-      // 日付の範囲でフィルタ
-      filtered = filtered.filter((row) => {
-        const date = new Date(row["aggregate from"]);
-        return date >= startDate && date <= endDate;
-      });
-
-      // 祝日を事前に取得しMap化
-      const holidays = holidayJP.between(startDate, endDate);
-      const holidayMap = new Map<string, string>();
-      holidays.forEach((h) => {
-        holidayMap.set(formatDate(h.date, "-"), h.name);
-      });
-
-      // 日ごとに集計
-      const dailyMap = new Map<string, AggregatedData>();
-      filtered.forEach((row) => {
-        const date = new Date(row["aggregate from"]);
-        const dayKey = formatDate(date, "-");
-        if (!dailyMap.has(dayKey)) {
-          const dayOfWeek = ["日", "月", "火", "水", "木", "金", "土"][date.getDay()];
-          const holidayName = holidayMap.get(dayKey) ?? "";
-          dailyMap.set(dayKey, {
-            ...row,
-            aggregateFrom: dayKey,
-            aggregateTo: dayKey,
-            totalCount: Number(row[TOTAL_COUNT_KEY]),
-            dayOfWeek,
-            holidayName,
-          });
-        }
-      });
-      setFilteredData(Array.from(dailyMap.values()));
-      return;
+      filtered = aggregateDaily(filtered, startDate, endDate);
     }
 
     // TODO:他の期間の処理を実装する
