+++ packages/shared/package.json
@@ -23,6 +23,7 @@
   },
   "dependencies": {
     "@fukui-kanko/shared": "workspace:^",
+    "@holiday-jp/holiday_jp": "^2.4.0",
     "@primer/octicons-react": "^19.14.0",
     "@radix-ui/react-label": "^2.1.1",
     "@radix-ui/react-popover": "^1.1.5",
@@ -43,18 +44,18 @@
     "tailwindcss": "^4.1.11"
   },
   "devDependencies": {
+    "@eslint/js": "^9.32.0",
+    "@types/papaparse": "^5.3.16",
     "@types/react": "^18.3.18",
     "@types/react-dom": "^18.3.5",
     "@vitejs/plugin-react": "^4.3.4",
-    "tw-animate-css": "^1.3.5",
-    "@eslint/js": "^9.32.0",
     "eslint": "^9.32.0",
     "eslint-plugin-react-hooks": "^5.2.0",
     "eslint-plugin-react-refresh": "^0.4.20",
     "globals": "^16.3.0",
-    "typescript-eslint": "^8.39.0",
-    "@types/papaparse": "^5.3.16",
+    "tw-animate-css": "^1.3.5",
     "typescript": "~5.6.2",
+    "typescript-eslint": "^8.39.0",
     "vite": "^6.0.5"
   }
 }
