+++ packages/shared/src/utils/aggregation.ts
@@ -1,0 +1,113 @@
+import * as holidayJP from "@holiday-jp/holiday_jp";
+import { AggregatedData, TOTAL_COUNT_KEY } from "../types";
+import { formatDate } from "./utils";
+
+function filterByRange(data: AggregatedData[], from: Date, to: Date) {
+  return data.filter((row) => {
+    const date = new Date(row["aggregate from"]);
+    return date >= from && date <= to;
+  });
+}
+
+/**
+ * 指定した期間内のデータを月単位で集計
+ */
+export function aggregateMonthly(data: AggregatedData[], start: Date, end: Date): AggregatedData[] {
+  const filtered = filterByRange(data, start, end);
+  const monthlyMap = new Map<string, AggregatedData>();
+  filtered.forEach((row) => {
+    const date = new Date(row["aggregate from"]);
+    const monthKey = `${date.getFullYear()}-${String(date.getMonth() + 1).padStart(2, "0")}`;
+    if (!monthlyMap.has(monthKey)) {
+      monthlyMap.set(monthKey, {
+        ...row,
+        aggregateFrom: `${monthKey}`,
+        aggregateTo: `${monthKey}`,
+        totalCount: Number(row[TOTAL_COUNT_KEY]),
+      });
+    } else {
+      const prev = monthlyMap.get(monthKey)!;
+      monthlyMap.set(monthKey, {
+        ...prev,
+        totalCount: Number(prev.totalCount) + Number(row[TOTAL_COUNT_KEY]),
+      });
+    }
+  });
+  return Array.from(monthlyMap.values());
+}
+
+/**
+ * 指定した期間内のデータを週単位で集計
+ */
+export function aggregateWeekly(
+  data: AggregatedData[],
+  startWeekRange: { from: Date; to: Date },
+  endWeekRange: { from: Date; to: Date },
+): AggregatedData[] {
+  const filtered = filterByRange(data, startWeekRange.from, endWeekRange.to);
+
+  const weeklyAggregated: AggregatedData[] = [];
+  let i = 0;
+  let isFirstWeek = true;
+  while (i < filtered.length) {
+    let weekRows: AggregatedData[] = [];
+    if (isFirstWeek) {
+      // 最初の週だけ他の週と範囲が異なる場合があるためfilterで抽出(例：2024/10/17からの週)
+      weekRows = filtered.filter((row) => {
+        const d = new Date(row["aggregate from"]);
+        return d >= startWeekRange.from && d <= startWeekRange.to;
+      });
+      // 件数分インデックスを進める
+      i += weekRows.length;
+      isFirstWeek = false;
+    } else {
+      // 以降は7日ごと
+      weekRows = filtered.slice(i, i + 7);
+      i += 7;
+    }
+    const total = weekRows.reduce((sum, row) => sum + Number(row[TOTAL_COUNT_KEY]), 0);
+    weeklyAggregated.push({
+      ...weekRows[0],
+      aggregateFrom: `${formatDate(new Date(weekRows[0]["aggregate from"]))}〜`,
+      aggregateTo: `${formatDate(new Date(weekRows[weekRows.length - 1]["aggregate from"]))}`,
+      totalCount: total,
+    });
+  }
+  return weeklyAggregated;
+}
+
+/**
+ * 指定した期間内のデータを日単位で集計（曜日・祝日名付き）
+ */
+export function aggregateDaily(data: AggregatedData[], start: Date, end: Date): AggregatedData[] {
+  const filtered = filterByRange(data, start, end);
+
+  // 祝日を事前に取得しMap化
+  const holidays = holidayJP.between(start, end);
+  const holidayMap = new Map<string, string>();
+  holidays.forEach((h) => {
+    holidayMap.set(formatDate(h.date, "-"), h.name);
+  });
+
+  // 日ごとに集計
+  const dailyMap = new Map<string, AggregatedData>();
+  filtered.forEach((row) => {
+    const date = new Date(row["aggregate from"]);
+    const dayKey = formatDate(date, "-");
+    const dayOfWeek = ["日", "月", "火", "水", "木", "金", "土"][date.getDay()];
+    const holidayName = holidayMap.get(dayKey) ?? "";
+
+    if (!dailyMap.has(dayKey)) {
+      dailyMap.set(dayKey, {
+        ...row,
+        aggregateFrom: dayKey,
+        aggregateTo: dayKey,
+        totalCount: Number(row[TOTAL_COUNT_KEY]),
+        dayOfWeek,
+        holidayName,
+      });
+    }
+  });
+
+  return Array.from(dailyMap.values());
+}
