+++ packages/shared/src/utils/index.ts
@@ -1,5 +1,6 @@
 // 共有ユーティリティ関数をここにエクスポート
 // 例: export { formatDate } from './dateUtils';
+export * from "./aggregation";
 export * from "./date";
 export * from "./getData";
 export * from "./utils";
